A public roadmap tool, based on Zendesk community
This is a rather extensive example of a rich app integrated into
Help Center. It provides a backend service which maps community
posts to a public-facing roadmap. The service has a set of REST
endpoints for both viewing and managing the roadmap.
A frontend app (written in React) can be included in a Custom Page
of Help Center, and be used to show the roadmap. Admins will also
be able to edit the roadmap from this app.
The app uses the HC integration token to authenticate the traffic.
The backend also has a test suite written in Postman to verify the
service's behavior.